Implement commands.ShowCommand.
Implement the commands.ShowCommand class. This is similar to the
commands.ListCommand class, and represents the toucan show command.
It resolves the patterns given to the command using a NameResolver,
and then is supposed to render the resulting objects to the standard
output using rendering.ShowRenderer.

Until rendering.ShowRenderer is implemented, the command simply prints
the number of objects it wants to show detailed info for onto the
standard output.