Edit schema and example setup file to work with toucan show.
toucan show needed some small changes to the schema in order for its full
functionality to be implemented. This is because both the schema and setup
files are works in progress as far as I can tell.

More work will be needed on the schema in future in order for 'reasons'
to work properly with toucan show, once the way of showing work-items has
been decided.