Add scenario tests for toucan show.
These tests are intended to validate if the right about of entries of each
type exist in the output (so if you try to show a lane, a name, description,
list of cards, list of views is shown).

In the future, it will probably be worthwhile adding tests to check that output
is correct (ie checking that the lane output shows the *correct* name), however
I would like to do that alongwith a bunch of other tests in its own patch once
toucan add, move and delete are implemented.