Make attachment path in setup files refer to the attachment file
This makes more sense than having it refer to the directory the
attachment is located in. The previous solution split up the
attachment path into "path" (directory) and "name" (file
basename). Just allowing "name" to be arbitrary and "path" to
specify to the attachment file is more logical and avoids
confusion.

This commit also allows attachment paths to be relative (to the
setup file), avoiding a hard-coded attachment path in the example
setup file.